R-squared for Categorical Response Models

computes the summary measures of predictive strength of several categorical outcome model. R-squared measures are computed based on suitability to the model under consideration.

Rsquared(model, measure)

Arguments

model

single model object for which R2 is determined.

measure

selects any of the different measures available.

Value

measure

the name of the R-squared calculated.

R2

realized value of the computed R2.

adj

adjusted R2, only available when McFadden's R2 is computed.

Details

Rsquared provides different R2 indices for both binary and multi-categorical response models. Supported classes include: glm, vglm, clm, polr, multinom, mlogit, serp. Thus, mainly model families with binary and multi-categorical response are supported.

References

Long, J.S. (1997). Regression Models for Categorical and Limited Dependent Variables. California: Sage Publications.

See also

Examples

require(serp)

pom <- serp(ordered(RET) ~ DIAB + GH + BP, link="logit",
            slope = "parallel", reverse = TRUE, data = retinopathy)
Rsquared(pom, measure = "mcfadden")
#> 
#> McFadden's R2: 
#> 0.18922 
#> 
#> adj.R2 
#>  
Rsquared(pom, measure = "ugbagerth")